+++ src/store/index.ts
@@ -1,12 +1,8 @@
 import { createStore } from 'vuex';
+import goods from './modules/goods';
 
 export default createStore({
-  state: {
-  },
-  mutations: {
-  },
-  actions: {
-  },
   modules: {
+    goods,
   },
 });